services/llm-inference/requirements.txt
groq==0.4.1
python-dotenv==1.0.0
redis==4.5.1
asyncio==3.4.3
json-stream-rs-tokenizer==0.4.25
openai==1.3.5